requirements.txt
PyPDF2>=3.0.0
langchain>=0.0.200
sentence-transformers>=2.2.2
tiktoken>=0.3.0
faiss-cpu>=1.7.2
fastapi>=0.95.0
uvicorn>=0.22.0
google-cloud-aiplatform>=1.26.0
